AG-17870 Keep tooltip code behind the Tooltip module
The tooltip refactor moved logic out of TooltipService into the call
sites, which pulled ~12kB of tooltip code into the core bundle for
grids that never register the Tooltip module.

Route it back through tooltipSvc using the existing "calls through to
standalone method for treeshaking" pattern, keeping the standalone
helpers as separate files:

- agColumn, cellCtrl and the header ctrls call tooltipSvc rather than
  importing the helpers directly
- cellTooltip.ts moves from rendering/cell to tooltip/
- full-width row tooltip state moves to tooltip/fullWidthRowTooltips.ts
- header tooltip setup returns the feature so the ctrls keep their
  tooltipFeature and componentTooltip fields private

Core bundle with no modules registered: 652.14kB -> 642.90kB raw
(183.71kB -> 181.37kB gzip). TooltipModule self size rises by the same
amount, with base+Tooltip unchanged.
