+++ packages/ag-grid-community/src/entities/agColumn.ts
@@ -23,7 +23,6 @@ import type {
 import type { IFrameworkEventListenerService } from '../interfaces/iFrameworkEventListenerService';
 import type { IRowNode } from '../interfaces/iRowNode';
 import type { SortDef, SortDirection, SortType } from '../interfaces/iSort';
-import { _isCellTooltipConfigured } from '../tooltip/tooltipValueUtils';
 import { _mergedEqual } from '../utils/mergeDeep';
 import { _clamp } from '../utils/number';
 import type { AgColumnGroup } from './agColumnGroup';
@@ -360,12 +359,7 @@ export class AgColumn<TValue = any>
     }
 
     private initTooltip(): void {
-        const colDef = this.colDef;
-        this.tooltipEnabled =
-            colDef.tooltip !== false &&
-            (_isCellTooltipConfigured(colDef) ||
-                colDef.tooltipComponent != null ||
-                colDef.tooltipComponentSelector != null);
+        this.beans.tooltipSvc?.initCol(this);
     }
 
     /** Kept apart from `resetActualWidth`, which `sizeColumnsToFit` also uses and must not change ownership. */
