+++ packages/ag-grid-community/src/rendering/row/fullWidthRowFeature.ts
@@ -1,6 +1,5 @@
 import { _findNextFocusableElement, _getActiveDomElement, _isBrowserSafari, _isFocusableFormField } from 'ag-stack';
 
-import { ComponentInstanceGuard } from '../../components/framework/componentInstanceGuard';
 import {
     _getFullWidthCellRendererDetails,
     _getFullWidthDetailCellRendererDetails,
@@ -9,8 +8,6 @@ import {
 } from '../../components/framework/userCompUtils';
 import { BeanStub } from '../../context/beanStub';
 import type { AgColumn } from '../../entities/agColumn';
-import type { ColDef } from '../../entities/colDef';
-import type { RowNode } from '../../entities/rowNode';
 import type { CellFocusedEvent } from '../../events';
 import { _addGridCommonParams } from '../../gridOptionsUtils';
 import type { CellPosition } from '../../interfaces/iCellPosition';
@@ -19,14 +16,7 @@ import type { ColumnPinnedType } from '../../interfaces/iColumn';
 import type { WithoutGridCommon } from '../../interfaces/iCommon';
 import type { UserCompDetails } from '../../interfaces/iUserCompDetails';
 import type { INotesFeature } from '../../interfaces/notes';
-import type { TooltipCallbackParams } from '../../tooltip/tooltipComponent';
-import type { TooltipFeature, TooltipSource, TooltipSourceParams } from '../../tooltip/tooltipFeature';
-import { _getCellTooltipComponentDefinition } from '../../tooltip/tooltipFeature';
-import {
-    _getLegacyTooltipFieldValue,
-    _isCellTooltipConfigured,
-    _resolveGroupTooltipValue,
-} from '../../tooltip/tooltipValueUtils';
+import type { FullWidthRowTooltips } from '../../tooltip/fullWidthRowTooltips';
 import { _isStopPropagationForAgGrid } from '../../utils/gridEvent';
 import type { CellCtrl } from '../cell/cellCtrl';
 import type { ICellRenderer, ICellRendererParams } from '../cellRenderers/iCellRenderer';
@@ -35,8 +25,7 @@ import type { FullWidthTarget, IRowModeFeature } from './iRowModeFeature';
 import type { RowCtrl } from './rowCtrl';
 
 export class FullWidthRowFeature extends BeanStub implements IRowModeFeature {
-    private readonly tooltipFeatures = new Map<HTMLElement, TooltipFeature>();
-    private readonly rendererGuards = new Map<HTMLElement, ComponentInstanceGuard>();
+    private tooltips: FullWidthRowTooltips | undefined;
     private notesFeature: INotesFeature | undefined;
     private focusEventWhileNotReady: CellFocusedEvent | null = null;
 
@@ -51,7 +40,7 @@ export class FullWidthRowFeature extends BeanStub implements IRowModeFeature {
             return;
         }
 
-        this.pruneStaleTooltipElements();
+        this.tooltips?.pruneStaleElements();
 
         if (this.shouldCreateCellSections() && rowComp.showEmbeddedFullWidth) {
             this.rowCtrl.isEmbeddedFullWidth = true;
@@ -89,7 +78,7 @@ export class FullWidthRowFeature extends BeanStub implements IRowModeFeature {
             return true;
         }
 
-        this.pruneStaleTooltipElements();
+        this.tooltips?.pruneStaleElements();
 
         if (this.shouldCreateCellSections() && rowComp.refreshEmbeddedFullWidth) {
             return rowComp.refreshEmbeddedFullWidth((pinned) => {
@@ -142,7 +131,8 @@ export class FullWidthRowFeature extends BeanStub implements IRowModeFeature {
         const { rowCtrl } = this;
         const { gos } = this;
         const { rowNode } = rowCtrl;
-        const rendererClaim = this.claimRenderer(eRow, adopt);
+        const tooltips = this.getTooltips();
+        const rendererClaim = tooltips?.claimRenderer(eRow, adopt);
         const params = _addGridCommonParams<ICellRendererParams>(gos, {
             fullWidth: true,
             data: rowNode.data,
@@ -158,8 +148,8 @@ export class FullWidthRowFeature extends BeanStub implements IRowModeFeature {
                 this.addFullWidthRowDragging(rowDraggerElement, dragStartPixels, value, rowDragEntireRow),
             setTooltip: (value, shouldDisplayTooltip) => {
                 gos.assertModuleRegistered('Tooltip', 3);
-                if (this.isAlive() && rendererClaim.isCurrent()) {
-                    this.setRendererTooltip(eRow, rowNode, value, shouldDisplayTooltip);
+                if (rendererClaim) {
+                    tooltips!.setRendererTooltip(eRow, rowNode, value, shouldDisplayTooltip, rendererClaim);
                 }
             },
         } as WithoutGridCommon<ICellRendererParams>);
@@ -176,7 +166,7 @@ export class FullWidthRowFeature extends BeanStub implements IRowModeFeature {
                 });
                 params.value = value;
                 params.valueFormatted = valueFormatted;
-                this.setupGroupRowsTooltip(rowNode, eRow);
+                tooltips?.setupGroupRowsTooltip(rowNode, eRow);
                 return _getFullWidthGroupCellRendererDetails(compFactory, params)!;
             }
             case 'FullWidthLoading':
@@ -193,203 +183,9 @@ export class FullWidthRowFeature extends BeanStub implements IRowModeFeature {
         this.beans.masterDetailSvc?.setupDetailRowAutoHeight(this.rowCtrl, eDetailGui);
     }
 
-    private setupFullWidthRowTooltip(
-        eGui: HTMLElement,
-        getTooltipValue: () => any,
-        shouldDisplayTooltip?: () => boolean,
-        getAdditionalParams?: () => TooltipSourceParams,
-        getTooltipComponentDefinition: () => ColDef | undefined = () => undefined
-    ) {
-        const tooltipSvc = this.beans.tooltipSvc;
-        if (!tooltipSvc || !this.rowCtrl.getCurrentRowElement()) {
-            return;
-        }
-
-        const source: TooltipSource = {
-            getGui: () => eGui,
-            getTooltipComponentDefinition,
-            getTooltipValue,
-            getLocation: () => 'fullWidthRow',
-            shouldDisplayTooltip,
-            ...(getAdditionalParams ? { getAdditionalParams } : {}),
-        };
-        const feature = tooltipSvc.registerTooltip(this, source, this.tooltipFeatures.get(eGui));
-        if (feature) {
-            this.tooltipFeatures.set(eGui, feature);
-        } else {
-            this.tooltipFeatures.delete(eGui);
-        }
-    }
-
-    private setRendererTooltip(
-        eGui: HTMLElement,
-        rowNode: RowNode,
-        value: string,
-        shouldDisplayTooltip?: () => boolean
-    ): void {
-        const isGroupRow = this.rowCtrl.getRowType() === 'FullWidthGroup';
-        if (value == null) {
-            this.clearTooltip(eGui);
-            if (isGroupRow) {
-                this.setupGroupRowsTooltip(rowNode, eGui);
-            }
-            return;
-        }
-        // Renderer-supplied tooltips are not gated by `tooltip: false`, so pass the colDef ungated.
-        const groupCol = isGroupRow ? (rowNode.rowGroupColumn as AgColumn | undefined) : undefined;
-        const colDef = isGroupRow ? this.getGroupColDef(rowNode) : undefined;
-        this.setupFullWidthRowTooltip(
-            eGui,
-            () => value,
-            shouldDisplayTooltip,
-            () => ({
-                colDef,
-                column: groupCol,
-                rowIndex: rowNode.rowIndex ?? 0,
-                node: rowNode,
-                data: rowNode.data,
-            }),
-            () => colDef
-        );
-    }
-
-    private clearTooltip(eGui: HTMLElement): void {
-        const feature = this.tooltipFeatures.get(eGui);
-        if (feature) {
-            this.destroyBean(feature, this.beans.context);
-        }
-        this.tooltipFeatures.delete(eGui);
-    }
-
-    /**
-     * `adopt` claims for a freshly mounted renderer, superseding earlier ones. A refresh claims
-     * provisionally: the surviving renderer's earlier params must keep working when it is kept in place.
-     */
-    private claimRenderer(eGui: HTMLElement, adopt: boolean) {
-        let guard = this.rendererGuards.get(eGui);
-        if (!guard) {
-            guard = new ComponentInstanceGuard();
-            this.rendererGuards.set(eGui, guard);
-        }
-        return adopt ? guard.claim() : guard.provisionalClaim();
-    }
-
-    /** Drop tooltip state keyed to elements the row no longer renders into (e.g. after a remount). */
-    private pruneStaleTooltipElements(): void {
-        const { tooltipFeatures, rendererGuards, rowCtrl } = this;
-        if (tooltipFeatures.size === 0 && rendererGuards.size === 0) {
-            return;
-        }
-
-        const rowComp = rowCtrl.getCurrentRowComp();
-        const currentElements = new Set<HTMLElement | undefined>([
-            rowCtrl.getCurrentRowElement() ?? undefined,
-            rowComp?.getPinnedLeftRowElement(),
-            rowComp?.getScrollingRowElement(),
-            rowComp?.getPinnedRightRowElement(),
-        ]);
-
-        for (const [element, feature] of tooltipFeatures) {
-            if (!currentElements.has(element)) {
-                this.destroyBean(feature, this.beans.context);
-                tooltipFeatures.delete(element);
-            }
-        }
-        for (const [element, guard] of rendererGuards) {
-            if (!currentElements.has(element)) {
-                guard.invalidate();
-                rendererGuards.delete(element);
-            }
-        }
-    }
-
-    /**
-     * Regular row grouping: read tooltip config from the row-group column's colDef.
-     * Tree data (no rowGroupColumn): fall back to the auto-group column def.
-     */
-    private getGroupColDef(rowNode: RowNode): ColDef | undefined {
-        const groupCol = rowNode.rowGroupColumn as AgColumn | undefined;
-        return groupCol?.colDef ?? this.gos.get('autoGroupColumnDef');
-    }
-
-    /** Resolves a full-width group tooltip lazily from its owning group column or tree-data auto column. */
-    private setupGroupRowsTooltip(rowNode: RowNode, eGui: HTMLElement): void {
-        const groupCol = rowNode.rowGroupColumn as AgColumn | undefined;
-        const { gos } = this;
-
-        const colDef = this.getGroupColDef(rowNode);
-        if (!colDef) {
-            this.clearTooltip(eGui);
-            return;
-        }
-
-        const hasTooltipValue = _isCellTooltipConfigured(colDef);
-        const hasLegacyComponentOnlyTooltip =
-            colDef.tooltip !== false &&
-            !hasTooltipValue &&
-            !!(colDef.tooltipComponent || colDef.tooltipComponentSelector);
-        if (!hasTooltipValue && !hasLegacyComponentOnlyTooltip) {
-            this.clearTooltip(eGui);
-            return;
-        }
-
-        const { valueSvc } = this.beans;
-        gos.assertModuleRegistered('Tooltip', 3);
-
-        let latestValueFormatted: string | null | undefined;
-        const getDisplay = () => {
-            const display = valueSvc.getValueForDisplay({ node: rowNode, includeValueFormatted: true, from: 'edit' });
-            latestValueFormatted = display.valueFormatted;
-            return display;
-        };
-
-        this.setupFullWidthRowTooltip(
-            eGui,
-            () => {
-                const { value, valueFormatted } = getDisplay();
-                const callbackParams = _addGridCommonParams<TooltipCallbackParams>(gos, {
-                    location: 'fullWidthRow' as const,
-                    colDef,
-                    column: groupCol,
-                    rowIndex: rowNode.rowIndex ?? 0,
-                    node: rowNode,
-                    data: rowNode.data,
-                    value,
-                    valueFormatted: valueFormatted ?? undefined,
-                });
-                if (hasTooltipValue) {
-                    return _resolveGroupTooltipValue(colDef, callbackParams, () => {
-                        const tooltipField = colDef.tooltipField;
-                        if (!tooltipField) {
-                            return { resolved: false };
-                        }
-                        const data = rowNode.data;
-                        if (!data) {
-                            // Regular row-grouping group nodes carry no `data`; use the group display value.
-                            return { resolved: true, value };
-                        }
-                        const containsDots = groupCol
-                            ? groupCol.tooltipFieldContainsDots
-                            : !gos.get('suppressFieldDotNotation') && tooltipField.includes('.');
-                        return {
-                            resolved: true,
-                            value: _getLegacyTooltipFieldValue(data, tooltipField, containsDots),
-                        };
-                    });
-                }
-                return value;
-            },
-            undefined,
-            () => ({
-                colDef,
-                column: groupCol,
-                rowIndex: rowNode.rowIndex ?? 0,
-                node: rowNode,
-                data: rowNode.data,
-                valueFormatted: latestValueFormatted ?? undefined,
-            }),
-            () => _getCellTooltipComponentDefinition(colDef)
-        );
+    /** Lazily created so a grid without the Tooltip module never instantiates tooltip state. */
+    private getTooltips(): FullWidthRowTooltips | undefined {
+        return (this.tooltips ??= this.beans.tooltipSvc?.createFullWidthRowTooltips(this.rowCtrl));
     }
 
     private addFullWidthRowDragging(
@@ -720,15 +516,7 @@ export class FullWidthRowFeature extends BeanStub implements IRowModeFeature {
     }
 
     public override destroy(): void {
-        const { context } = this.beans;
-        for (const feature of this.tooltipFeatures.values()) {
-            this.destroyBean(feature, context);
-        }
-        this.tooltipFeatures.clear();
-        for (const guard of this.rendererGuards.values()) {
-            guard.invalidate();
-        }
-        this.rendererGuards.clear();
+        this.tooltips = this.destroyBean(this.tooltips, this.beans.context);
         this.notesFeature?.destroy();
         this.notesFeature = undefined;
         super.destroy();
