+++ packages/ag-grid-community/src/tooltip/headerTooltipSource.ts
@@ -5,12 +5,14 @@ import type { AgColumn } from '../entities/agColumn';
 import type { AgColumnGroup } from '../entities/agColumnGroup';
 import type { ColDef, ColGroupDef } from '../entities/colDef';
 import { _addGridCommonParams } from '../gridOptionsUtils';
+import type { HeaderCellCtrl } from '../headerRendering/cells/column/headerCellCtrl';
+import type { HeaderGroupCellCtrl } from '../headerRendering/cells/columnGroup/headerGroupCellCtrl';
 import type { TooltipCallbackParams } from './tooltipComponent';
-import type { TooltipSource } from './tooltipFeature';
+import type { TooltipFeature, TooltipSource } from './tooltipFeature';
 import { _getHeaderTooltipComponentDefinition, _isShowTooltipWhenTruncated } from './tooltipFeature';
 import { _resolveHeaderTooltipValue } from './tooltipValueUtils';
 
-interface ComponentTooltip {
+export interface ComponentTooltip {
     value?: any;
     shouldDisplay?: () => boolean;
 }
@@ -27,8 +29,7 @@ interface HeaderTooltipSourceParams {
     hasCustomHeader: () => boolean;
 }
 
-/** @internal AG_GRID_INTERNAL - Not for public use. Can change / be removed at any time. */
-export function _createHeaderTooltipSource(params: HeaderTooltipSourceParams): TooltipSource {
+function createHeaderTooltipSource(params: HeaderTooltipSourceParams): TooltipSource {
     const { beans, eGui, location, column, getColDef, getComponentTooltip, getDisplayName } = params;
     let latestDisplayName: string | null | undefined;
     const isOverflowing = _isElementOverflowingCallback(
@@ -75,4 +76,54 @@ export function _createHeaderTooltipSource(params: HeaderTooltipSourceParams): T
             };
         },
     };
+}
+
+export function _setupHeaderTooltip(
+    beans: BeanCollection,
+    ctrl: HeaderCellCtrl,
+    current: TooltipFeature | undefined,
+    getComponentTooltip: () => ComponentTooltip
+): TooltipFeature | undefined {
+    if (!ctrl.isAlive()) {
+        return current;
+    }
+    const { column, eGui } = ctrl;
+
+    const source = createHeaderTooltipSource({
+        beans,
+        eGui,
+        location: 'header',
+        column,
+        getColDef: () => column.colDef,
+        getComponentTooltip,
+        getDisplayName: () => beans.colNames.getDisplayNameForColumn(column, 'header', true),
+        overflowElementSelector: '.ag-header-cell-text',
+        hasCustomHeader: () => !!column.colDef.headerComponent,
+    });
+    return beans.tooltipSvc!.registerTooltip(ctrl, source, current);
+}
+
+export function _setupHeaderGroupTooltip(
+    beans: BeanCollection,
+    ctrl: HeaderGroupCellCtrl,
+    current: TooltipFeature | undefined,
+    getComponentTooltip: () => ComponentTooltip
+): TooltipFeature | undefined {
+    if (!ctrl.isAlive()) {
+        return current;
+    }
+    const { column, eGui } = ctrl;
+
+    const source = createHeaderTooltipSource({
+        beans,
+        eGui,
+        location: 'headerGroup',
+        column,
+        getColDef: () => column.getColGroupDef(),
+        getComponentTooltip,
+        getDisplayName: () => beans.colNames.getDisplayNameForColumnGroup(column, 'header'),
+        overflowElementSelector: '.ag-header-group-text',
+        hasCustomHeader: () => !!column.getColGroupDef()?.headerGroupComponent,
+    });
+    return beans.tooltipSvc!.registerTooltip(ctrl, source, current);
 }
